fix(tabs): prevent wrong activeTab index (#271)
## 📃 Summary

closes https://github.com/shortcuts/no-neck-pain.nvim/issues/268 (again)

we were relying on a default (and wrong) initial `activeTab` value which
was set to 1. While working properly when enabling on the tab 1, then
the others, it broke when we were initializing the plugin at a tab index
other than 1.

we now rely on an helper than ensures the `activeTab` comes from the
neovim state and not hardcoded.