feat: provide pre/post enable and pre/post disable callback methods (#451)
## 📃 Summary

This PR is a follow up of
https://github.com/shortcuts/no-neck-pain.nvim/pull/448, closes
https://github.com/shortcuts/no-neck-pain.nvim/issues/111

----

This is for the feature mentioned in #111, or at least some of it: User
can specify callbacks in the config, that will run before and after the
buffer is centered or restored. Nothing happens by default.
Callbacks take no arguments, because I don't know what values exist that
would be useful to the user.

closes https://github.com/shortcuts/no-neck-pain.nvim/issues/111

## 📸 Preview
For example, to disable the individual status line, and hide the command
line, but restore both options when toggling this plugin off:
```lua
    require("no-neck-pain").setup({
    -- Some options ...
        callbacks = {
            preEnable = function()
                vim.opt.laststatus = 0
                vim.opt.cmdheight  = 0
            end,
            postDisable = function()
                vim.opt.laststatus = 2
                vim.opt.cmdheight  = 1
            end
        },
    -- Some more options ...
})
```

## Notes:
* I wouldn't know how to add tests for this feature, and maybe there is
a better way of doing this. I just made a pull request because this
feature was the thing I missed the most.

* [Goyo.vim](https://github.com/junegunn/goyo.vim) does this by
providing `GoyoEnter` and `GoyoLeave` autocommands. I suppose there is
an advantage in being able to define autocmd's at any point, but I can't
think of anything that can't be defined during the initial
configuration.

* Thank you for making this plugin. It is the only text-centering plugin
that didn't require me to change other parts of my config.

---------

Co-authored-by: Fabrice-Bernes <Fabrice.bernes@protonmail.com>